Split the Models module into DataModel and QmlModel
