+++ geometry_constructor/Application.py
@@ -1,6 +1,6 @@
 import sys
 from os import path
-from geometry_constructor.Models import InstrumentModel
+from geometry_constructor.QmlModel import InstrumentModel
 from geometry_constructor.Writers import HdfWriter, Logger
 from PySide2.QtCore import QUrl, QObject
 from PySide2.QtQml import QQmlApplicationEngine, qmlRegisterType
