+++ tests/test_instrument_model.py
@@ -1,23 +1,24 @@
-from geometry_constructor import Models
+from geometry_constructor import DataModel
+from geometry_constructor.QmlModel import InstrumentModel
 
 
 def test_initialise_model():
-    model = Models.InstrumentModel()
+    model = InstrumentModel()
     assert model.rowCount() == 1
-    assert isinstance(model.components[0], Models.Sample)
+    assert isinstance(model.components[0], DataModel.Sample)
 
 
 def test_add_component():
-    model = Models.InstrumentModel()
+    model = InstrumentModel()
     model.add_detector('My Detector')
     assert model.rowCount() == 2
-    assert isinstance(model.components[1], Models.Detector)
+    assert isinstance(model.components[1], DataModel.Detector)
     assert model.components[1].name == 'My Detector'
 
 
 def test_remove_component():
-    model = Models.InstrumentModel()
+    model = InstrumentModel()
     model.add_detector('My Detector')
     model.remove_component(1)
     assert model.rowCount() == 1
-    assert not isinstance(model.components[0], Models.Detector)
+    assert not isinstance(model.components[0], DataModel.Detector)
